Separate shared types into their own submodule
